[FIX] 이용약관 API Settings 도메인으로 옮기기 + API 버전 오기입 수정
