+++ src/main/java/com/moongeul/backend/api/setting/controller/SettingController.java
@@ -1,7 +1,9 @@
 package com.moongeul.backend.api.setting.controller;
 
+import com.moongeul.backend.api.setting.dto.AgreeTermsRequestDTO;
 import com.moongeul.backend.api.setting.dto.PrivacyLevelResponseDTO;
 import com.moongeul.backend.api.setting.dto.PrivacyLevelUpdateRequestDTO;
+import com.moongeul.backend.api.setting.service.AgreeTermsSettingService;
 import com.moongeul.backend.api.setting.service.PrivacySettingService;
 import com.moongeul.backend.common.response.ApiResponse;
 import com.moongeul.backend.common.response.SuccessStatus;
@@ -13,19 +15,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @Tag(name = "Setting", description = "설정 관련 API 입니다.")
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/v1/setting")
+@RequestMapping("/api/v2/setting")
 public class SettingController {
 
     private final PrivacySettingService privacySettingService;
+    private final AgreeTermsSettingService agreeTermsSettingService;
 
     @Operation(
             summary = "계정 공개 범위 조회 API",
@@ -67,5 +66,21 @@ public class SettingController {
 
         PrivacyLevelResponseDTO response = privacySettingService.updatePrivacyLevel(userDetails.getUsername(), request);
         return ApiResponse.success(SuccessStatus.UPDATE_PRIVACY_LEVEL_SUCCESS, response);
+    }
+
+    @Operation(
+            summary = "이용약관동의 API",
+            description = "이용약관 동의 여부를 저장합니다."
+    )
+    @ApiResponses({
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "닉네임 중복 체크 성공")
+    })
+    @PostMapping("/agree-terms")
+    public ResponseEntity<ApiResponse<Void>> agreeToTerms(
+            @AuthenticationPrincipal UserDetails userDetails,
+            @RequestBody AgreeTermsRequestDTO agreeTermsRequestDTO) {
+
+        agreeTermsSettingService.agreeToTerms(userDetails.getUsername(), agreeTermsRequestDTO);
+        return ApiResponse.success_only(SuccessStatus.TERMS_AGREE_SUCCESS);
     }
 }
