+++ src/main/java/com/moongeul/backend/api/setting/entity/TermsType.java
@@ -1,4 +1,4 @@
-package com.moongeul.backend.api.member.entity;
+package com.moongeul.backend.api.setting.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
