+++ src/main/java/com/moongeul/backend/api/setting/repository/AgreeRepository.java
@@ -1,8 +1,8 @@
-package com.moongeul.backend.api.member.repository;
+package com.moongeul.backend.api.setting.repository;
 
-import com.moongeul.backend.api.member.entity.Agree;
+import com.moongeul.backend.api.setting.entity.Agree;
 import com.moongeul.backend.api.member.entity.Member;
-import com.moongeul.backend.api.member.entity.Terms;
+import com.moongeul.backend.api.setting.entity.Terms;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
